Fixed UI and organized code
