+++ src/app/List_a_cause/See_Preview/Success/page.tsx
@@ -32,7 +32,6 @@ export default function SuccessPage() {
   }, []);
 
   // Define the cause URL to share.
-  // Adjust the URL pattern below to point to your cause page.
   const causeUrl =
     typeof window !== "undefined" && formData
       ? `${window.location.origin}/List_a_cause/${encodeURIComponent(
@@ -110,7 +109,7 @@ export default function SuccessPage() {
     const diffTime = deadlineDate.getTime() - now.getTime();
     const diffDays = Math.ceil(diffTime / (1000 * 60 * 60 * 24));
     if (diffDays <= 0) return "Past due";
-    // For numbers 10 or below, convert to words (optional)
+    // Convert numbers 10 or below to words (optional)
     const words = [
       "zero",
       "one",
@@ -127,81 +126,104 @@ export default function SuccessPage() {
     const dayWord = diffDays <= 10 ? words[diffDays] : diffDays.toString();
     return `${dayWord} day${diffDays > 1 ? "s" : ""} left`;
   }
-  
+
+  // Track whether the bottom panel is open or collapsed
+  const [isOpen, setIsOpen] = useState(false);
 
   return (
-    <div>
-      <div className="flex flex-col items-center justify-center text-center space-y-4">
-        <div className="text-xl font-semibold">
+    <div className="min-h-screen bg-gray-50 pb-24">
+      {/* Top Section: Congratulatory Message */}
+      <div className="px-4 py-8 flex flex-col items-center justify-center text-center space-y-4">
+        <h1 className="text-2xl font-semibold">
           🎉 Congratulations! Your Cause is Live 🎉
-        </div>
-        <div className="text-gray-600">
+        </h1>
+        <p className="text-gray-600 text-sm md:text-base max-w-lg">
           Thank you for sharing your cause with RefreeG. Your cause is now live
-          on our platform, <br />
-          ready to reach supporters who care about making a difference. Let&apos;s
-          get the word out!
+          on our platform, <br /> ready to reach supporters who care about
+          making a difference. Let&lsquo;s get the word out!
+        </p>
+        <div className="py-6">
+          <Image
+            src="/list_a_cause/success/success.svg"
+            width={250}
+            height={250}
+            alt="Successful submission illustration"
+            className="mx-auto"
+          />
         </div>
-        <Image
-          src="/list_a_cause/success/success.svg"
-          width={250}
-          height={250}
-          alt="Successful submission illustration"
-          className="py-6"
-        />
       </div>
 
-      <div className="bg-[#214570] text-white px-3 pt-3 rounded-t-3xl fixed bottom-0 left-0 w-full z-50">
-        <Image
-          src="/list_a_cause/success/toggleBar.png"
-          width={50}
-          height={10}
-          alt="Toggle bar"
-          className="mx-auto mb-3 w-24"
-        />
+      {/* Bottom Panel with "Pull-up" Effect */}
+      <div
+        className={`
+          fixed bottom-0 left-0 w-full z-50 bg-[#214570] text-white px-4 pt-4 pb-6 rounded-t-3xl
+          transition-transform duration-300
+          ${isOpen ? "translate-y-0" : "translate-y-[calc(100%-80px)]"}
+        `}
+      >
+        {/* Toggle Handle */}
+        <div className="flex justify-center">
+          <Image
+            src="/list_a_cause/success/toggleBar.png"
+            width={50}
+            height={10}
+            alt="Toggle bar"
+            className="mb-3 w-24 cursor-pointer"
+            onClick={() => setIsOpen(!isOpen)}
+          />
+        </div>
 
-        <div className="flex justify-between">
-          <div className="sideA w-1/2">
-            <div className="flex mb-3">
-              {formData && formData.uploadedImage ? (
-                <Image
-                  src={formData.uploadedImage.src}
-                  alt={formData.uploadedImage.name}
-                  width={50}
-                  height={40}
-                  className="w-48"
-                />
-              ) : (
-                <Image
-                  src="/list_a_cause/success/attachedImage.svg"
-                  width={50}
-                  height={40}
-                  alt="Attached Image"
-                  className="w-48"
-                />
-              )}
-              <div className="pl-2">
-                <div className="text-lg">
+        <div className="flex flex-col md:flex-row justify-between items-start md:items-center">
+          {/* Side A: Cause Details */}
+          <div className="w-full md:w-1/2 mb-4 md:mb-0">
+            <div className="flex items-center space-x-3">
+              <div className="w-16 h-16 flex-shrink-0">
+                {formData && formData.uploadedImage ? (
+                  <Image
+                    src={formData.uploadedImage.src}
+                    alt={formData.uploadedImage.name}
+                    width={64}
+                    height={64}
+                    className="rounded object-cover"
+                  />
+                ) : (
+                  <Image
+                    src="/list_a_cause/success/attachedImage.svg"
+                    width={64}
+                    height={64}
+                    alt="Attached Image"
+                    className="rounded object-cover"
+                  />
+                )}
+              </div>
+              <div className="text-left">
+                <h2 className="text-lg font-medium">
                   {formData ? formData.causeTitle : "Your Cause Title"}
-                </div>
-                <div className="text-xs">
+                </h2>
+                <p className="text-xs">
                   Goal Amount: {formData ? formData.goalAmount : "N/A"}
-                </div>
-                <div className="text-xs">
+                </p>
+                <p className="text-xs">
                   Category: {formData ? formData.causeCategory : "N/A"}
-                </div>
-                <div className="text-xs">
-  Deadline:{" "}
-  {formData?.deadline
-    ? `${formData.deadline} | ${getDaysLeft(formData.deadline)}`
-    : "N/A"}
-</div>
-
+                </p>
+                <p className="text-xs">
+                  Deadline:{" "}
+                  {formData?.deadline
+                    ? `${formData.deadline} | ${getDaysLeft(formData.deadline)}`
+                    : "N/A"}
+                </p>
               </div>
             </div>
-            <div className="mb-2 underline">What&apos;s next?</div>
-            <div className="mb-2 underline">Support and FAQs</div>
+            <div className="mt-3 space-y-1 text-left">
+              <button className="underline text-sm block">
+                What&apos;s next?
+              </button>
+              <button className="underline text-sm block">
+                Support and FAQs
+              </button>
+            </div>
             <button
-              className="flex items-center space-x-2 bg-white text-black py-2 px-8 mb-8 rounded-md"
+              className="mt-4 flex items-center justify-center space-x-2 bg-white text-black py-2 px-4 rounded-md w-full"
               aria-label="View cause page"
             >
               <span>View cause page</span>
@@ -214,95 +236,77 @@ export default function SuccessPage() {
             </button>
           </div>
 
-          <div className="sideB flex w-1/2">
-            <div className="pr-2">
-              <Image
-                src="/list_a_cause/success/line.svg"
-                alt=""
-                width={1}
-                height={2}
-              />
-            </div>
-            <div>
-              <div className="pb-4">
-                Help your cause gain visibility! Share it on social media to
-                reach more potential supporters.
-              </div>
-              <div className="flex pb-4">
-                <div className="pr-4">
-                  <a onClick={handleCopyLink} style={{ cursor: "pointer" }}>
-                    <Image
-                      src="/list_a_cause/success/copyLink.svg"
-                      alt="Copy link"
-                      width={30}
-                      height={30}
-                    />
-                  </a>
-                </div>
-                <div className="pr-4">
-                  <a onClick={handleFacebookShare} style={{ cursor: "pointer" }}>
-                    <Image
-                      src="/list_a_cause/success/facebook.svg"
-                      alt="Facebook"
-                      width={30}
-                      height={30}
-                    />
-                  </a>
-                </div>
-                <div className="pr-4">
-                  <a onClick={handleInstagramShare} style={{ cursor: "pointer" }}>
-                    <Image
-                      src="/list_a_cause/success/instagram.svg"
-                      alt="Instagram"
-                      width={30}
-                      height={30}
-                    />
-                  </a>
-                </div>
-                <div className="pr-4">
-                  <a onClick={handleYouTubeShare} style={{ cursor: "pointer" }}>
-                    <Image
-                      src="/list_a_cause/success/youtube.svg"
-                      alt="YouTube"
-                      width={30}
-                      height={30}
-                    />
-                  </a>
-                </div>
-                <div className="pr-4">
-                  <a onClick={handleTwitterShare} style={{ cursor: "pointer" }}>
-                    <Image
-                      src="/list_a_cause/success/x.svg"
-                      alt="Twitter"
-                      width={30}
-                      height={30}
-                    />
-                  </a>
-                </div>
-                <div className="pr-4">
-                  <a onClick={handleWhatsAppShare} style={{ cursor: "pointer" }}>
-                    <Image
-                      src="/list_a_cause/success/whatsapp.svg"
-                      alt="WhatsApp"
-                      width={30}
-                      height={30}
-                    />
-                  </a>
-                </div>
-              </div>
-              <button
-                className="flex items-center space-x-2 bg-white text-black py-2 px-8 rounded-md"
-                aria-label="Go to dashboard"
-              >
-                <span>Go to dashboard</span>
+          {/* Divider (Only on medium screens and up) */}
+          <div className="hidden md:block w-px bg-white mx-4" />
+
+          {/* Side B: Social Sharing */}
+          <div className="w-full md:w-1/2">
+            <p className="mb-4 text-sm">
+              Help your cause gain visibility! Share it on social media to reach
+              more potential supporters.
+            </p>
+            <div className="flex flex-wrap items-center gap-4 mb-4">
+              <button onClick={handleCopyLink} className="cursor-pointer">
                 <Image
-                  src="/list_a_cause/success/chevronRight.svg"
-                  alt="Right arrow"
-                  width={15}
-                  height={15}
+                  src="/list_a_cause/success/copyLink.svg"
+                  alt="Copy link"
+                  width={30}
+                  height={30}
+                />
+              </button>
+              <button onClick={handleFacebookShare} className="cursor-pointer">
+                <Image
+                  src="/list_a_cause/success/facebook.svg"
+                  alt="Facebook"
+                  width={30}
+                  height={30}
+                />
+              </button>
+              <button onClick={handleInstagramShare} className="cursor-pointer">
+                <Image
+                  src="/list_a_cause/success/instagram.svg"
+                  alt="Instagram"
+                  width={30}
+                  height={30}
+                />
+              </button>
+              <button onClick={handleYouTubeShare} className="cursor-pointer">
+                <Image
+                  src="/list_a_cause/success/youtube.svg"
+                  alt="YouTube"
+                  width={30}
+                  height={30}
+                />
+              </button>
+              <button onClick={handleTwitterShare} className="cursor-pointer">
+                <Image
+                  src="/list_a_cause/success/x.svg"
+                  alt="Twitter"
+                  width={30}
+                  height={30}
+                />
+              </button>
+              <button onClick={handleWhatsAppShare} className="cursor-pointer">
+                <Image
+                  src="/list_a_cause/success/whatsapp.svg"
+                  alt="WhatsApp"
+                  width={30}
+                  height={30}
                 />
               </button>
             </div>
+            <button
+              className="flex items-center justify-center space-x-2 bg-white text-black py-2 px-4 rounded-md w-full"
+              aria-label="Go to dashboard"
+            >
+              <span>Go to dashboard</span>
+              <Image
+                src="/list_a_cause/success/chevronRight.svg"
+                alt="Right arrow"
+                width={15}
+                height={15}
+              />
+            </button>
           </div>
         </div>
       </div>
