+++ src/components/ui/navbar.tsx
@@ -103,7 +103,7 @@ export function Navbar({ userSession }: { userSession?: boolean }) {
           </DropdownMenu>
         </MenuLink>
 
-        <MenuLink href="#" className="text-white hover:text-white bg-blue-600 hover:bg-blue-700">
+        <MenuLink href="/List_a_cause" className="text-white hover:text-white bg-blue-600 hover:bg-blue-700">
           List a cause
         </MenuLink>
